Take class hierarchy into account when resolving
Both when resolving handlers and creators, the hierarchy of classes was not interpreted correctly. This has been resolved.

In addition, I discovered that the current interceptor implementation stood in the way of this fix, and that it should be implemented in another way. Interceptors are not handlers, after all, but this mechanism was abused for that. As such, the CommandHandlerInterceptor support has been removed. This may be re-added in the future.