Fix false positive on versions not equal to 4
The version check was quite primitive, checking the jar name and extracting the name from that. This made it quite unreliable, and disabled the plugin for users with their own `axon-*` module (#327).

The version check has now been changed. JARs found to contain the name axon will now be checked its META-INF for it's groupId and artifactID, and only be disabled/detected if it is the correct group.

Resolves #327